requirement.txt
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
asttokens==2.4.1
blis==0.7.11
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
dill==0.3.8
distro==1.9.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.2
executing==2.0.1
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
graphviz==0.20.3
grpcio==1.65.4
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
icecream==2.1.3
idna==3.7
importlib_resources==6.4.0
isort==5.13.2
Jinja2==3.1.4
jsonpickle==3.2.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mccabe==0.7.0
mdurl==0.1.2
murmurhash==1.0.10
networkx==3.2.1
numpy==1.26.4
openai==1.38.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
preshed==3.0.9
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pylint==3.2.6
pyparsing==3.1.2
pytextrank==3.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
rich==13.7.1
rsa==4.9
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.5
tomli==2.0.1
tomlkit==0.13.0
tqdm==4.66.5
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
zipp==3.19.2
